+++ app/src/main/java/com/todobom/opennotescanner/SettingsActivity.java
@@ -1,0 +1,22 @@
+package com.todobom.opennotescanner;
+
+import android.app.FragmentManager;
+import android.app.FragmentTransaction;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class SettingsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_settings);
+        FragmentManager fm=getFragmentManager();
+        FragmentTransaction ft=fm.beginTransaction();
+
+        SettingsFragment sf=new SettingsFragment();
+        ft.replace(android.R.id.content, sf);
+        ft.commit();
+    }
+
+}